Задание #17 Первый класс
